refactor(SaveAsImage): consolidate files
