+++ client/src/dialogs/DialogRoot.jsx
@@ -7,7 +7,7 @@ import AboutDialog from './About'
 import AnalyticsDialog from './Analytics'
 import FeatureFlagDialog from './FeatureFlagDialog'
 import GeotagDialog from './Geotag'
-import SaveAsImageDialog from './SaveAsImageDialog'
+import SaveAsImageDialog from './SaveAsImage'
 import SettingsDialog from './SettingsDialog'
 import SignInDialog from './SignInDialog'
 import WhatsNewDialog from './WhatsNewDialog'
