+++ client/src/dialogs/SaveAsImage/SaveAsImageDialog.scss
@@ -60,9 +60,6 @@
 }
 
 .save-as-image-preview-image {
-  /* Removes extra baseline space */
-  line-height: 0;
-
   /* Adds a grid for transparent backgrounds */
   background-image: linear-gradient(45deg, #eee 25%, transparent 25%),
     linear-gradient(-45deg, #eee 25%, transparent 25%),
@@ -75,12 +72,14 @@
     12px -12px,
     -12px 0;
 
+  /* Display below border */
+  z-index: -1;
+
   img {
     /* Images may have variable dimensions. Use max-width/max-height
     to preserve aspect ratio */
     max-width: 100%;
     max-height: 300px;
-    cursor: move; // IE
     cursor: grab;
 
     /* Center image within preview container */
